增加注释 调整enums变量顺序 抽出一个线程模块 数据库暂时改为了单例模式 修改部分路径的合并方式 抽出signals模块，解决循环依赖的问题 修改control文件，兼容18.04 取消一个多重判断使用，方便添加特殊软件包 删除不必要的文件夹，增加ignore文件，初步优化部分方法
