+++ .gitignore
@@ -1,0 +1,4 @@
+/.idea
+/.idea.moved
+/.vscode
+/.pybuild
