--- .idea/.name
@@ -1,1 +1,0 @@
-ubuntu-kylin-software-center
