--- .idea/misc.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.6.8 (/usr/bin/python3.6)" project-jdk-type="Python SDK" />
-</project>
-
