--- .idea/modules.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/ubuntu-kylin-software-center.iml" filepath="$PROJECT_DIR$/.idea/ubuntu-kylin-software-center.iml" />
-    </modules>
-  </component>
-</project>
-
