+++ debian/control
@@ -2,7 +2,7 @@ Source: ubuntu-kylin-software-center
 Section: misc
 Priority: optional
 Maintainer: Ubuntu Kylin Team <ubuntukylin-members@list.launchpad.net>
-Build-Depends: debhelper (>= 12),
+Build-Depends: debhelper (>= 9),
                dh-python,
                python3-dev,
                python3-distutils-extra,
