+++ kydroid/downloadmanager.py
@@ -28,9 +28,10 @@ import stat
 import shutil
 from urllib import request
 from PyQt5.QtCore import *
-from models.enums import KYDROID_SOURCE_SERVER, KYDROID_DOWNLOAD_PATH, Signals, AppActions
+from models.enums import KYDROID_SOURCE_SERVER, KYDROID_DOWNLOAD_PATH,  AppActions
 import threading
 from models.globals import Globals
+from models.signals import Signals 
 
 def download_sourcelist():
     slurl = os.path.join(KYDROID_SOURCE_SERVER, 'kydroid-sourcelist')
