+++ ui/configwidget.py.old
@@ -26,7 +26,7 @@ from PyQt5.QtWidgets import *
 from PyQt5.QtGui import *
 from PyQt5.QtCore import *
 from ui.confw import Ui_ConfigWidget
-from models.enums import Signals
+#from models.signals import Signals 
 
 
 class ConfigWidget(QWidget):
