+++ ui/taskwidget.py
@@ -27,7 +27,7 @@ import sip
 from PyQt5.QtWidgets import *
 from PyQt5.QtGui import *
 from PyQt5.QtCore import *
-from models.enums import Signals
+from models.signals import Signals 
 
 #
 # class Taskwidget(QWidget,Signals):
